requirements.txt
Django
djangorestframework
Pillow
psycopg2
django-filter
djangorestframework-simplejwt
